Added JWT auth

+++ graphql/resolvers/user/deleteUserResolver.js
@@ -1,5 +1,8 @@
 const models = require("../../../models");
-module.exports = (source, { id }) => {
+module.exports = (source, { id }, {tokenPayload}) => {
+  if(!tokenPayload) {
+    return null;
+  }
   return models.User.destroy({
     where: {
       id
